requirements.txt
# ─── CORE DJANGO + REST ────────────────────────────────────────────────────────
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1

# ─── AUTENTICACIÓN Y OAUTH2 ────────────────────────────────────────────────────
dj-rest-auth==7.0.1
django-allauth==65.7.0
django-oauth-toolkit==3.0.1
oauthlib==3.2.2
jwcrypto==1.5.6
google-auth==2.40.3

# ─── BASE DE DATOS MySQL ───────────────────────────────────────────────────────
mysqlclient==2.2.7

# ─── UTILIDADES ────────────────────────────────────────────────────────────────
pillow==11.1.0
requests==2.32.3
drf-yasg==1.21.10

# ─── WEBSOCKETS / ASGI ─────────────────────────────────────────────────────────
channels==4.1.0
daphne==4.1.0

# ─── CLOUDINARY STORAGE ────────────────────────────────────────────────────────
cloudinary==1.44.0
django-cloudinary-storage==0.3.0

# ─── SERVIR ARCHIVOS ESTÁTICOS ─────────────────────────────────────────────────
whitenoise==6.9.0

# ─── PAGOS PAYPAL ──────────────────────────────────────────────────────────────
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1

# ─── GENERACIÓN DE PDF ─────────────────────────────────────────────────────────
xhtml2pdf==0.2.17
reportlab==4.4.1
